requirements.txt
# Essential dependencies for AI Video Workflow
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pydantic>=2.0.0

# Google Cloud dependencies for Veo 3.1
google-cloud-storage>=2.0.0
google-cloud>=0.34.0

# Video processing dependencies
moviepy>=1.0.3

# Streamlit UI dependencies
streamlit>=1.28.0

# Optional: LangChain integration (uncomment if needed)
# langchain>=0.1.0
# langchain-google-genai>=1.0.0